Création trait et remplacement du code répété par ces traits dans computer, device et component
